+++ app/styles/_components.scss
@@ -6,4 +6,4 @@
 @import '../components/grid/base';
 @import '../components/icon-label-button/base';
 @import '../components/product-item/base';
-@import '../components/notification/base';
+@import '../components/notification-manager/base';
